use float as it works for int or float

+++ scripts/split_qscore.py
@@ -137,9 +137,9 @@ def find_score(read, ftype):
         header = read[0].split(" ")
         for i in header:
             if i[:4] == "qs:i":
-                score = int(i[5:])
+                score = float(i[5:])
             elif i.split("=")[0] in ["qscore", "mean_qscore"]:
-                score = int(i.split("=")[1])
+                score = float(i.split("=")[1])
             else:
                 score = None
     
@@ -147,7 +147,7 @@ def find_score(read, ftype):
         sread = read.split("\t")
         for i in sread:
             if i[:4] == "qs:i":
-                score = int(i[5:])
+                score = float(i[5:])
             else:
                 score = None
     return score
